backend/pyproject.toml
[tool.poetry]
name = "App"
version = "0.1.0"
description = ""
authors = ["Evin"]
packages = [{include = "App"}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
asyncpg = "^0.30.0"
alembic = "^1.16.1"
aiogram = "^3.20.0.post0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"